\docType{data}
\name{inventarioPackages}
\alias{inventarioPackages}
\title{Inventario de paquetes}
\format{
Data frame con inventario de paquetes utilizados en el curso.
}
\usage{
inventarioPackages
}
\description{
Base de datos con inventario de paquetes utilizados en el curso.
}
\examples{
data(inventarioPackages)
}
\keyword{datasets}
