% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pruebas_app.R
\name{run_pruebas_app}
\alias{run_pruebas_app}
\title{Ejecutar App Shiny de Pruebas de Hipotesis}
\usage{
run_pruebas_app()
}
\value{
Inicia la aplicacion Shiny.
}
\description{
Lanza una app Shiny con pruebas de hipotesis tradicionales y resumidas.
}
