% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata_size.R
\name{strata_size}
\alias{strata_size}
\title{Distribucion muestral para muestreo estratificado}
\usage{
strata_size(samp_size, porcentajes, estratos = NULL)
}
\arguments{
\item{samp_size}{tamano total de la muestra.}

\item{porcentajes}{vector numerico con los porcentajes de los estratos.
Puede estar en escala 0-100 o 0-1.}

\item{estratos}{vector opcional con los nombres de los estratos.}
}
\value{
Lista con la informacion suministrada y la distribucion muestral por estrato.
}
\description{
Distribuye un tamano de muestra total entre estratos usando los porcentajes
suministrados para cada estrato.
}
\examples{
strata_size(385, c(40, 35, 25), c("Norte", "Centro", "Sur"))
strata_size(100, c(0.5, 0.3, 0.2))
}
