% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.prop.R
\name{test.prop}
\alias{test.prop}
\title{Prueba z de hipotesis para proporciones con datos resumidos}
\usage{
test.prop(
  x,
  n,
  p = NULL,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95
)
}
\arguments{
\item{x}{numero de exitos. Para dos proporciones, vector de longitud 2.}

\item{n}{tamano de muestra. Para dos proporciones, vector de longitud 2.}

\item{p}{proporcion bajo H0 en la prueba de una proporcion. Para dos
proporciones se usa diferencia nula igual a 0.}

\item{alternative}{tipo de hipotesis alternativa: "two.sided", "less" o "greater".}

\item{conf.level}{nivel de confianza para el intervalo de confianza.}
}
\value{
Objeto clase \code{htest}, con estadistico z, p-value, intervalo de
confianza en \code{conf.int} y alias \code{int.conf}, estimacion e hipotesis alternativa.
}
\description{
Realiza una prueba z para una proporcion o para la diferencia de dos
proporciones usando el numero de exitos y el tamano de muestra.

Esta prueba esta pensada para muestras grandes. Por eso exige que todos los
tamanos de muestra sean mayores que 30.
}
\examples{
res <- test.prop(x = 45, n = 100, p = 0.5)
res$int.conf

res2 <- test.prop(x = c(56, 42), n = c(120, 110))
res2$int.conf
}
