% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted_size.R
\name{adjusted_size}
\alias{adjusted_size}
\title{Correccion del tamano de la muestra por el factor de poblacion finita}
\usage{
adjusted_size(samp_size, pob_size)
}
\arguments{
\item{samp_size}{valor del tamano de la muestra inicial}

\item{pob_size}{valor del tamano de la poblacion}
}
\value{
n0 tamano de muestra ajustado
}
\description{
Funcion que corrige el tamano de la muestra cuando n/N>0.05
}
\examples{
 adjusted_size(385,500)
}
